requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.9
django-allauth==0.35.0
django-cors-headers==2.2.0
django-rest-auth==0.9.3
django-webpack-loader==0.6.0
djangorestframework==3.7.7
drfdocs==0.0.11
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.7
psycopg2==2.7.3.2
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
